requirements.txt
numpy>=1.21.0
pandas>=1.3.0
seaborn>=0.11.0
matplotlib>=3.4.0
beautifulsoup4>=4.9.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0
google-auth>=2.0.0
oauth2client>=4.1.3
python-dotenv>=0.19.0
requests>=2.26.0
tenacity>=8.0.0
backoff>=2.2.0
pytz>=2021.1
python-dateutil>=2.8.2 
certifi>=2021.5.30 